Add Algorithms analyser

+++ src/main/kotlin/com/okhhamed/apiratelimiter/limiter/RateLimiterManager.kt
@@ -6,7 +6,9 @@ import com.okhhamed.apiratelimiter.config.RateLimiterConfig
 class RateLimiterManager {
     private var rateLimiter : RateLimiter = when(RateLimiterConfig.enabledAlgorithm) {
         AlgorithmType.TOKEN_BUCKET -> TokenBucketRateLimiter(RateLimiterConfig.TOKEN_BUCKET_CAPACITY, RateLimiterConfig.TOKEN_BUCKET_REFILL_RATE)
-        AlgorithmType.LEAKY_BUCKET -> LeakyBucketRateLimiter(RateLimiterConfig.LEAKY_BUCKET_CAPACITY, RateLimiterConfig.LEAKY_BUCKET_LEAK_RATE)
+        AlgorithmType.LEAKY_BUCKET -> LeakyBucketRateLimiter(RateLimiterConfig.LEAKY_BUCKET_CAPACITY,
+            RateLimiterConfig.LEAKY_BUCKET_LEAK_RATE.toInt()
+        )
         AlgorithmType.FIXED_WINDOW -> TODO("Implement Fixed Window")
         AlgorithmType.SLIDING_WINDOW -> SlidingWindowRateLimiter(RateLimiterConfig.MAX_REQUESTS, RateLimiterConfig.WINDOW_IN_MILLS)
     }
